Fix issues with Neko tool accessibility and mesh generation (#36)
This pull request addresses two issues. Firstly, it ensures that the
Neko tool is accessible from the examples. Secondly, it runs mesh
generation in the preparation step. The changes include updating the
mesh file path and adding using a Neko tool to generate a mesh box.
Additionally, the initializer in the simcomp module has been moved.