Fix job submission and improve logging (#43)
This pull request includes several commits that fix issues related to
job submission and improve logging and messages to the user. The changes
ensure that the job submission works correctly on different clusters,
such as MareNostrum 5 and DTU. The submission process has been updated
to use batch scripts associated with the desired cluster, and the output
of the job is stored in the results folder after completion.
Additionally, the logging and messages to the user have been improved
for better clarity and understanding.