masks (#48)
Here I've added masks based on `point_zones` for both the design
indicator and source terms.
- volume constraint is now based on the optimization domain
- we have some basic masking functionality in `mask_ops` but I intend to
extend this in the future for more complex functionality (and GPUs)
- this is a big ***warning*** for Abbas (currently implementing the
`optimizer_t`) I've just masked the sensitivity. This is not the correct
way of handling the interface with MMA, because the $N$ is wrong. You
should use the masks on initialization of MMA such that only those
within the mask are flagged as design variables.