Update documentation (#54)
The pull request includes several commits that update and improve the
documentation. The changes involve moving things back to the
documentation, cleaning up the documentation, and updating the table of
contents to render correctly. Additionally, the pull request ensures
that the documentation is not broken on pull requests and deploys the
documentation on the develop branch instead of the main branch.