Test working on a merge queue (#80)
A few updates to the CI. Some of this is to support the merge queues by
github.
The merge queue ensure multiple PR's being accepted at roughly the same
time does not break the repository. Each accepted Pr is added to a queue
and merged in order. This should help streamline our development, while
every merge is tested before the next is added.

- [x] Develop should properly support the merge queues.
- [x] The formatting action is now functional.
- [x] Compilation check should fail on warnings, such as un-initialized
variables.