Documentation for adjoint fluid + name changes (#99)
The PR is moving in the direction of getting the adjoint passive scalar
off the ground. It consists of:
1) A bit of documentation for the adjoint velocity equations (I'll fill
in the adjoint passive scalar in a future PR). It isn't linked to the
rest of the docs, I'll wait for PR#95.
2) A lot of name changes such that `adjoint` becomes `adjoint_fluid`
where applicable, opening the way for us to have
- `fluid` and `adjoint_fluid`
- `scalar` and `adjoint_scalar`