Mapping cascade (#124)
This PR has a couple of nice features, but is still a bit of a work in
progress (maybe)

- The main contribution is the notion of a mapping cascade and a lot of
JSON support for the pre-existing mappings. Now we prescribe a list of
mappings which are performed sequentially, and it also takes care of the
chain rule moving backwards.
- There are some cosmetic changes to variable names, particularly
`sens_in` and `sens_out` :sweat_smile:
- Obviously much of the mapping was inspired by the source term handler,
so in principal this could be extended to allow user defined mappings in
some way, but this is to be discussed.

This PR is still in draft mode because I think we should really discuss
who owns the mapping, and how we should handle mapping to multiple
coeffients, and how we should hand constraints defined on intermediate
steps (ie, the potential for the volume constraint to defined on the
filtered design variable)

There are still a few cosmetics changes that need to be done so I'm
leaving this on draft mode, but take a look if you can and tell me what
you think

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>