Initial material distribution (#127)
This one is pretty self explanatory.

But let me know your thoughts on moving the boundary_mesh utilities out
of the `brinkman_source_term` and into `set_scalar_ic` in `neko`, and
then we all have public access to that functionality.

I guess something also up for discussion is how multiple point zones is
handled. Currently it's quite nice in the `brinkman_source_term` but I
also saw that `point_zones` support boolean operations, so maybe this is
the way forward.

https://neko.cfd/docs/develop/da/dd0/point-zones.html

Depends on #124

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>
Co-authored-by: Abbas Mousavi <smousavi@kth.se>
Co-authored-by: Tim Felle Olsen <timfelle@hotmail.com>