Adjustments for the mixer tutorial. (#158)
On Thursday 3rd of April I went through the tutorial with some other
`neko` colleagues at KTH. This PR contains some suggestions made during
that session.

*note* This is not a finalized set of parameters, that will come in a
separate PR.

Also, it was suggested by Martin that `user_velocity_pointwise` should
be switched to `user_dirichlette` as it may be depreciated in the
future. This will also come in a separate PR.