Documentation for objectives and constraints (#160)
depends on #159 and #158.

Continuing with the theme of documentation for the user guide, this one
fleshes out the objectives and constraints.

One thing I don't like so much about this PR is it is very topology
optimization of fluids specific. When the abstract MMA PR comes through
you need to remember to modify this documentation making everything here
a specific subsection of "topology optimization objectives and
constraints", or something to this effect.