Refactor/mma device (#163)
The aim of this PR was to refactor mma_device but it ended up being more
of a cleanup.
Reason is that I wanted to change the code such that we conduct small
calculations (for vectors of size m, number of constraints) on CPU and
only offload calculations over n (number of design variables) to the
device.
However, it turns out that in that case, we need to move data back and
forth between host and device too many times.
Therefore, I have decided to keep stuff as they were.

Long story short, just a little cleanup PR :))