Update MMA functions and ensure array compatibility (#184)
Refactor MMA function interfaces for improved clarity and ensure that
arrays are correctly initialized and updated based on device support.
This enhances the overall functionality and performance of the MMA
module.

Pure CPU and pure GPU works. Mixing does not, so the default is
currently what works.

Let's get this in to make