Post processing and documentation for the mixer (#197)
This PR contains
- a python script for plotting the convergence history (it's very
simple... maybe we should consider something a bit nicer or removing it
completely)
- A `.pvsm` file for post processing the mixer. It looks quite nice I
think! I included the university logos because I thought it would look
cool, but actually in my opinion it's a bit cluttered and I would vote
to remove them and only keep the CEEC logo. But we'll put it to vote in
the meeting on Wednesday.

![heavy](https://github.com/user-attachments/assets/0a326c61-441f-41d2-bd46-3d100c096fbe)

*** NOTE *** I've put the "help wanted" because maybe it's better to
keep the `.pvsm` in some other folder and I'm not so sure how well this
will play with `run.sh` on completion.

Maybe it should just be a part of the mixer tutorial.

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>