Introduce an additional coef of the GL space (#265)
This additional coef on the GL space is used for over integration.
This rather hacky right now needing to create a brand new dofmap, gs etc
just to get the coefficients.
I intend to modify things on the neko side allowing a coef to be
generated completely with a preexisting gs etc and a new space.

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>