Update the sensitivity driver (#272)
This was a big change from the original but the motivation was the
following: I wanted to test not just the objectives/constraints, but the
whole problem interface too. This lead to the switch from prescribing
various `.f90` files to various `.case` files.

Another big change is that Niels and Casper really want to see that
linear slope, not just sensitivity within some tolerance. So now we dump
a `.csv` file for each test, and they can be plotted with
`plotter_FD_check.py`.

I've put the help wanted flag on because I'm incredibly sus on the CMake
script, this is not my forte.

Another big consideration is that right now we can only do steady runs,
which means we have to wait a LONG time for these tests to complete.
However, Casper mention that if we're just interested in verifying the
adjoint, we can do that with an unsteady case and only a few timesteps,
which is a really good idea (we just don't have that functionality yet)
So ideally in the future these tests will become a lot faster.

depends on #243

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>