Add GPU support for sensitivity (#299)
This PR ensure's the CPU and GPU versions of #247 are consistent.

here is a comparison between CPU and GPU (note the number of elements
were slightly higher for the GPU version)

The key takeaways, the minimum dissipation is still slightly broken and
I've tried EVERYTHING. But the scalar (which in the end is the important
one) and lube term look excellent.

<img width="1484" height="792" alt="image"
src="https://github.com/user-attachments/assets/4669604b-7483-40b7-9099-301df339b19e"
/>

I've also found an indexing miss match in the masks which I remember
being very conscious of at the time, but it appears this inconsistency
was fixed on the neko side?

---------

Co-authored-by: Tim Felle Olsen <tife@dtu.dk>